refactor. 폴더 구조 정리
- components의 common 폴더 내에 공통으로 사용하는 component와 element가 혼재되어 있어 분리하였습니다.
