+++ src/App.tsx
@@ -5,7 +5,7 @@ import { ReactQueryDevtools } from 'react-query/devtools';
 import { Provider } from 'react-redux';
 import { RouterProvider } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
-import Loader from './components/common/Loader';
+import Loader from './elements/Loader';
 import store from './redux/store';
 import GlobalStyle from './shared/style/GlobalStyle';
 import theme from './shared/style/theme';
