+++ src/components/CartItem.tsx
@@ -1,5 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import Button from '../elements/Button';
 import useViewport from '../hooks/useViewport';
 import usePutCartQuery from '../query/usePutCartQuery';
 import type { TCartList } from '../redux/reducer/cartSlice';
@@ -13,7 +14,6 @@ import theme from '../shared/style/theme';
 import type { IProductPayCart, TUserOption } from '../shared/types/types';
 import * as t from '../style/cartItem.style';
 import CartOptionModal from './CartOptionModal';
-import Button from './common/Button';
 import GlobalModal from './common/GlobalModal';
 
 interface IProps {
