+++ src/components/CartOptionModal.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useMemo } from 'react';
 import toast from 'react-hot-toast';
 import { IoCloseOutline } from 'react-icons/io5';
+import Button from '../elements/Button';
 import useGetDetailQuery from '../query/useGetDetailQuery';
 import usePutCartQuery from '../query/usePutCartQuery';
 import { setOptions } from '../redux/reducer/optionSlice';
 import { useAppDispatch, useAppSelector } from '../redux/store';
 import theme from '../shared/style/theme';
 import * as t from '../style/cartOptionModal.style';
-import Button from './common/Button';
 import Option from './common/Option';
 
 type TProps = {
