+++ src/components/LoginSearchPwd.tsx
@@ -1,4 +1,7 @@
 import React, { useEffect, useReducer, useState } from 'react';
+import Button from '../elements/Button';
+import ErrMsg from '../elements/ErrorMsg';
+import Input from '../elements/Input';
 import type { TSearchPwd } from '../query/useGetLoginSearchPwdQuery';
 import { useGetLoginSearchPwdQuery } from '../query/useGetLoginSearchPwdQuery';
 import type { TNewPwd } from '../query/usePutLoginNewPwdQuery';
@@ -6,9 +9,6 @@ import usePutLoginNewPwdQuery from '../query/usePutLoginNewPwdQuery';
 import theme from '../shared/style/theme';
 import { NewPwdInitial } from '../shared/utils/inputInitialValue';
 import { PwdReducer } from '../shared/utils/inputReducer';
-import Button from './common/Button';
-import ErrMsg from './common/ErrorMsg';
-import Input from './common/Input';
 
 export default function LoginSearchPwd() {
   // 비밀번호 찾기
