+++ src/components/MyWithdrawalModal.tsx
@@ -1,7 +1,7 @@
+import Button from '../elements/Button';
 import useDeleteWithdrawalQuery from '../query/useDeleteWithdrawlQuery';
 import theme from '../shared/style/theme';
 import * as t from '../style/myWithdrawalModal.style';
-import Button from './common/Button';
 
 type TProps = {
   onClose: () => void;
