+++ src/components/PaymentAddDefault.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useReducer, useState } from 'react';
 import { shallowEqual } from 'react-redux';
+import Button from '../elements/Button';
+import ErrMsg from '../elements/ErrorMsg';
+import Input from '../elements/Input';
 import { setName, setPayInfo, setPhone } from '../redux/reducer/payInputSlice';
 import { useAppDispatch, useAppSelector } from '../redux/store';
 import theme from '../shared/style/theme';
 import { PayInputInitial } from '../shared/utils/inputInitialValue';
 import { PayReducer } from '../shared/utils/inputReducer';
 import * as t from '../style/paymentInput.style';
-import Button from './common/Button';
-import ErrMsg from './common/ErrorMsg';
-import Input from './common/Input';
 
 export type TUser = { d_Name: string; d_Phone: string };
 type TProps = {
