+++ src/components/PaymentAddNew.tsx
@@ -1,6 +1,9 @@
 import React, { useEffect, useReducer, useState } from 'react';
 import toast from 'react-hot-toast';
 import { shallowEqual } from 'react-redux';
+import Button from '../elements/Button';
+import ErrMsg from '../elements/ErrorMsg';
+import Input from '../elements/Input';
 import {
   setAddress1,
   setAddress2,
@@ -17,9 +20,6 @@ import { PayInputInitial } from '../shared/utils/inputInitialValue';
 import { PayReducer } from '../shared/utils/inputReducer';
 import * as t from '../style/paymentInput.style';
 import Address from './common/Address';
-import Button from './common/Button';
-import ErrMsg from './common/ErrorMsg';
-import Input from './common/Input';
 import type { TUser } from './PaymentAddDefault';
 
 type TProps = {
