+++ src/components/detail/DetailContents.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import useGetReviewsQuery from '../query/useGetReviewsQuery';
-import { delivery } from '../shared/utils/imgUrls';
-import * as t from '../style/detailContents.style';
+import useGetReviewsQuery from '../../query/useGetReviewsQuery';
+import { delivery } from '../../shared/utils/imgUrls';
+import * as t from '../../style/detailContents.style';
 import DetailReviews from './DetailReviews';
 
 function DetailContents({ productNo, content }: PropsType) {
