+++ src/pages/PaymentPage.tsx
@@ -2,12 +2,12 @@ import { useEffect, useMemo, useState } from 'react';
 import toast from 'react-hot-toast';
 import { shallowEqual } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router-dom';
-import Button from '../components/common/Button';
-import Input from '../components/common/Input';
 import PayDrop from '../components/common/PayDrop';
 import type { Tprops } from '../components/common/ProductCard';
 import ProductCard from '../components/common/ProductCard';
 import PaymentInput from '../components/PaymentInput';
+import Button from '../elements/Button';
+import Input from '../elements/Input';
 import useGetPaymentQuery from '../query/useGetPaymentQuery';
 import usePostPaymentQuery from '../query/usePostPaymentQuery';
 import { setPayInfo } from '../redux/reducer/payInputSlice';
