+++ src/pages/SignUpPage.tsx
@@ -2,10 +2,10 @@ import React, { useEffect, useReducer, useState } from 'react';
 import toast from 'react-hot-toast';
 import { useNavigate } from 'react-router-dom';
 import Address from '../components/common/Address';
-import Button from '../components/common/Button';
-import ErrMsg from '../components/common/ErrorMsg';
-import Input from '../components/common/Input';
 import Profile from '../components/common/Profile';
+import Button from '../elements/Button';
+import ErrMsg from '../elements/ErrorMsg';
+import Input from '../elements/Input';
 import type { TSignUpData } from '../query/usePostSignupQuery';
 import usePostSignupQuery from '../query/usePostSignupQuery';
 import theme from '../shared/style/theme';
